Expose Annotation::setFromPickResult and use iniside AnnotationsPlugin::createAnnotation
